cleanDestDir and refactorings
