+++ lib/app.js
@@ -25,11 +25,11 @@ class App {
         this.#srcDir      = config.sourceDir;
         this.#destDir     = config.destDir; // needed as attr?
         this.#zipPath     = config.zipPath;
-        this.#includes    = config.includes; // TODO make single include dir?
+        this.#includes    = config.includes;
         this.#logger      = config.logger;
         this.#withMeta    = config.withMeta;
 
-        this.#proveParamsAndDirs();
+        this.initDestDir({ cleanDestDirOnStart: config.cleanDestDirOnStart });
         this.#excelFileManager = this.#setupExcelFileManager(config.eol, config.skipFieldsStartingWith);
         this.#includeFileManager = this.#setupIncludeFileManager();
         this.#metaCalculator = this.#withMeta && new MetaCalculator({
@@ -44,11 +44,16 @@ class App {
         });
     }
 
-    #proveParamsAndDirs() {
-        // TODO check includedir! Or remove the check at all
-        // check destDir not empty
-        if (!fs.existsSync(this.#srcDir)) throw Error('Source dir does not exist');
-        if (this.#destDir && !fs.existsSync(this.#destDir)) fs.mkdirSync(this.#destDir);
+    initDestDir({cleanDestDirOnStart}) {
+        if (fs.existsSync(this.#destDir)) {
+            if (cleanDestDirOnStart) {
+                this.#logger.log(chalk.grey('Removing dest dir'), this.#destDir);
+                fs.rmdirSync(this.#destDir, { recursive: true, force: true });
+                fs.mkdirSync(this.#destDir);
+            }
+        } else {
+            fs.mkdirSync(this.#destDir);
+        }
     }
 
     #setupExcelFileManager(eol, skipFieldsStartingWith) {
@@ -70,7 +75,10 @@ class App {
 
     #setupIncludeFileManager() {
         if (!this.#includes || this.#includes.length === 0) return;
-        // support just one include dir for now ???
+        if (this.#includes.length > 1) {
+            throw Error('Multiple includes is not supported curently, please log an issue if you need this feature');
+        }
+
         const includeManager = new IncludeFileManager({
             includeDir: this.#includes[0],
             destDir: this.#destDir,
